Rename workflow to contribute (module, column)
