+++ template.html.jinja
@@ -9,7 +9,7 @@
 <thead>
 <tr>
   <th>language</th>
-  <th>workflow</th>
+  <th>contribute</th>
   <th>build</th>
   <th><a href="https://plausible.io/data-policy#how-we-count-unique-users-without-cookies" target="_blank">visitors</a></th>
   <th>translators</th>
@@ -28,7 +28,7 @@
   {% else %}
   <td data-label="language">{{ language }}</td>
   {% endif %}
-  <td data-label="workflow">
+      <td data-label="contribute">
     {% if contrib_link %}<a href="{{ contrib_link }}" target="_blank">{% endif %}
     {% if on_platform %}platform{% else %}repository{% endif %}
     {% if contrib_link %}</a>{% endif %}
